Selective Rendering added
- Renamed WorldTypes -> ActorTypes
- Rendering is now selective - rather than render every time of every
layer, the layers are combined and then compared with the previous
frame. Only differences with the previous frame are rendered. This kills
the need for thread sleeps and removes all flickering.

-Empty spaces are now identified by a new Empty class. This way, they
are not read in as nulls. Required for selective rendering to
differentiate between a new empty cell and a cell that hasn't changed.